Support Untagged Documents (#382)
* Fixes item search

* Finishes API work for docs

* Merge fix
