+++ api/swagger/swagger.yaml
@@ -5891,6 +5891,11 @@ paths:
           required: false
           description: "Query: multiple parameter instances are supported using OR."
           type: string
+        - in: query
+          name: categorized
+          required: false
+          description: "Query: indicating if a documents should be categorized or not"
+          type: boolean
       responses:
         "200":
           description: "Success"
